Tested basic app and added bounds on zoom and longitude
The testing went completely fine it all worked as expected. There is no simple
way to add bounds on these values. Instead the class now listens for camera
change events from the map and repositions the view afterwards. There was
initially a problem with this if the user exceeded two bounds in one movement
to fix this instead of animating for each bound that was exceeded a new
position is created based on all of the bounds and the current position. Then
the map is animated to this new position.