Revert "Started work on refactoring the code to follow model, view controller."
This reverts commit 23754406045eb60ab665e7ac1142792c9a26c702.

This was done because the code was resulting in an error message saying that
a file could not be accessed. Despite removing all of the file reading changes
this was still happening so instead it was decided to return to a time before
the error and work forwards carefully testing each stage.