Added basic file reading to the application.
Created a text file asset from which to read. This text file currently
contains only information needed for the view about Royal College.

Also created code to read the text file and interpret in a way to
recognise the buildings name and location.

To demonstrate this a shape was drawn on top of the royal college.