Improved navigation within the app.
Added navigational features:
Added bounds while "inside" a building - this ensures the app can keep
track of where a user is by forcing them to go back outside and into a new
building if they wish to view the rooms.
Added a memory to the controller to allow it to keep track of which floor
each building is on.