Major refactor brought all seperate apps together.
This should make it easier to use the seperate applications quickly. Also
made the outside links the same as all of the others. This eliminates the
need to consult googles directions api which was frequently taking a very
long time and occasionally crashing the entire app. It also frequently
gave worse drawn routes than simply adding them in the same way the indoor
routes are handled.